Avoid keeping business logic in __init__.py
