+++ src/pysonar_scanner/__main__.py
@@ -18,10 +18,11 @@
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 
-from pysonar_scanner import app_logging, cache
-from pysonar_scanner import configuration
+from pysonar_scanner import app_logging
+from pysonar_scanner import cache
 from pysonar_scanner.api import get_base_urls, SonarQubeApi
-from pysonar_scanner.configuration import ConfigurationLoader
+from pysonar_scanner.configuration import configuration_loader
+from pysonar_scanner.configuration.configuration_loader import ConfigurationLoader
 from pysonar_scanner.configuration.properties import SONAR_VERBOSE
 from pysonar_scanner.scannerengine import ScannerEngine
 
@@ -43,6 +44,6 @@ def set_logging_options(config):
 
 
 def __build_api(config: dict[str, any]) -> SonarQubeApi:
-    token = configuration.get_token(config)
+    token = configuration_loader.get_token(config)
     base_urls = get_base_urls(config)
     return SonarQubeApi(base_urls, token)
